Move StorageClient to own file
